Fix timeline position misalignment on project load
Fixes incorrect clip positioning that occurred when projects were saved
with one sample rate and loaded with another (or when sampleRate attribute
was missing/corrupted).

Changes:
- sproject.cpp: Add warning when sampleRate attribute is missing, making
  the issue visible to users. Preserves backwards compatibility.
- scut.cpp: Replace hard-coded 44100 Hz defaults with calculations based
  on actual project sample rate (3 locations: 2 constructors and XML load)
- slink.cpp: Use consistent integer type (unsigned long long) for timeline
  position serialization instead of converting through double
- smainwindow.cpp: Add missing QCloseEvent include and implement save-on-close

Timeline positions are stored in samples and scale with the project's
sample rate. When the rate was unknown/mismatched, all clips would shift
on the timeline. Now positions are preserved correctly across save/load.

+++ smaragd/main/src/slink.cpp
@@ -9,7 +9,7 @@ int SLink::serializeSelfAttributes( QTextStream &o )
     o << " objectId='" << reinterpret_cast<std::uintptr_t>(&object_) << "'"
       << " hasStartTime='" << hasStartTime() << "'";
     if( hasStartTime() ) {
-        o << " startTime='" << (double)getStartTime() << "'";
+        o << " startTime='" << (unsigned long long)getStartTime() << "'";
     }
     return 0;
 }
